+++ src/main/java/org/jabref/gui/journals/AbbreviationViewModel.java
@@ -10,63 +10,73 @@ import javafx.beans.property.StringProperty;
 import org.jabref.logic.journals.Abbreviation;
 
 /**
- * This class provides a view model for abbreviation objects which can also
- * define placeholder objects of abbreviations. This is indicated by using the
- * {@code pseudoAbbreviation} property.
+ * This class provides a view model for abbreviation objects which can also define placeholder objects of abbreviations.
+ * This is indicated by using the {@code pseudoAbbreviation} property.
  */
 public class AbbreviationViewModel {
 
     private final Abbreviation abbreviationObject;
     private final StringProperty name = new SimpleStringProperty("");
     private final StringProperty abbreviation = new SimpleStringProperty("");
+    private final StringProperty shortestUniqueAbbreviation = new SimpleStringProperty("");
     private final BooleanProperty pseudoAbbreviation = new SimpleBooleanProperty();
-
 
     public AbbreviationViewModel(Abbreviation abbreviation) {
         this.abbreviationObject = abbreviation;
-        pseudoAbbreviation.set(this.abbreviationObject == null);
+        this.pseudoAbbreviation.set(this.abbreviationObject == null);
         if (this.abbreviationObject != null) {
             this.name.bindBidirectional(this.abbreviationObject.nameProperty());
             this.abbreviation.bindBidirectional(this.abbreviationObject.abbreviationProperty());
-        } else {
-            this.name.set("Add new Abbreviation");
+            this.shortestUniqueAbbreviation.bindBidirectional(this.abbreviationObject.shortestUniqueAbbreviationProperty());
         }
     }
 
     public Abbreviation getAbbreviationObject() {
-        return this.abbreviationObject;
+        return abbreviationObject;
+    }
+
+    public String getName() {
+        return name.get();
     }
 
     public void setName(String name) {
         this.name.set(name);
     }
 
+    public String getAbbreviation() {
+        return abbreviation.get();
+    }
+
     public void setAbbreviation(String abbreviation) {
         this.abbreviation.set(abbreviation);
     }
 
-    public String getName() {
-        return this.name.get();
+    public String getShortestUniqueAbbreviation() {
+        return shortestUniqueAbbreviation.get();
     }
 
-    public String getAbbreviation() {
-        return this.abbreviation.get();
+    public void setShortestUniqueAbbreviation(String shortestUniqueAbbreviation) {
+        this.shortestUniqueAbbreviation.set(shortestUniqueAbbreviation);
     }
 
     public boolean isPseudoAbbreviation() {
-        return this.pseudoAbbreviation.get();
+        return pseudoAbbreviation.get();
     }
 
     public StringProperty nameProperty() {
-        return this.name;
+        return name;
     }
 
     public StringProperty abbreviationProperty() {
-        return this.abbreviation;
+        return abbreviation;
+    }
+
+    public StringProperty shortestUniqueAbbreviationProperty() {
+        return shortestUniqueAbbreviation;
     }
 
     public BooleanProperty isPseudoAbbreviationProperty() {
-        return this.pseudoAbbreviation;
+        return pseudoAbbreviation;
     }
 
     @Override